Add files via upload
Intravenous (IV) drip monitoring systems are used to monitor and regulate the delivery of fluids
and medication to patients through an IV line. The purpose of IV drip monitoring systems is to
ensure the safe and effective delivery of fluids and medications to patients in healthcare settings,
including hospitals, clinics, and home healthcare.